Bug 1165103 - Use a separate AppDelegate for Aurora
