+++ Client/Application/main.swift
@@ -2,6 +2,17 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-let isRunningTest = NSClassFromString("XCTestCase") != nil
-let appDelegate = isRunningTest ? TestAppDelegate.self : AppDelegate.self
+private var appDelegate: AppDelegate.Type
+
+if AppConstants.IsRunningTest {
+    appDelegate = TestAppDelegate.self
+} else {
+    switch AppConstants.BuildChannel {
+    case .Aurora:
+        appDelegate = AuroraAppDelegate.self
+    case .Developer:
+        appDelegate = AppDelegate.self
+    }
+}
+
 UIApplicationMain(Process.argc, Process.unsafeArgv, NSStringFromClass(UIApplication.self), NSStringFromClass(appDelegate))
